+++ NAI-Generator-Flutter/lib/data/models/param_config.dart
@@ -129,7 +129,7 @@ class ParamConfig {
       "noise_schedule": noiseSchedule,
       "skip_cfg_above_sigma": skipCfgAboveSigma,
       "use_coords": true,
-      "seed": randomSeed ? Random().nextInt(1 << 32 - 1) : seed,
+      "seed": randomSeed ? Random().nextInt((1 << 32) - 1) : seed,
       "characterPrompts": [],
       "v4_prompt": {},
       "v4_negative_prompt": {},